+++ logprep/ng/processor/clusterer/processor.py
@@ -44,7 +44,6 @@ Processor Configuration
 """
 
 import typing
-from typing import Tuple
 
 from attrs import define, field, validators
 
@@ -57,11 +56,11 @@ from logprep.processor.clusterer.signature_calculation.signature_phase import (
     SignaturePhaseStreaming,
 )
 from logprep.util.helper import (
+    MISSING,
+    FieldValue,
     add_fields_to,
     get_dotted_field_value,
     get_field_value_no_slice,
-    MISSING,
-    FieldValue,
 )